DDP-4276 move copy config classes to new package
