+++ pepper-apis/src/main/java/org/broadinstitute/ddp/db/dao/CopyConfigurationDao.java
@@ -13,12 +13,12 @@ import org.broadinstitute.ddp.db.DaoException;
 import org.broadinstitute.ddp.db.dto.QuestionDto;
 import org.broadinstitute.ddp.exception.OperationNotAllowedException;
 import org.broadinstitute.ddp.model.activity.types.QuestionType;
-import org.broadinstitute.ddp.model.event.CompositeCopyConfigurationPair;
-import org.broadinstitute.ddp.model.event.CopyAnswerLocation;
-import org.broadinstitute.ddp.model.event.CopyConfiguration;
-import org.broadinstitute.ddp.model.event.CopyConfigurationPair;
-import org.broadinstitute.ddp.model.event.CopyLocation;
-import org.broadinstitute.ddp.model.event.CopyLocationType;
+import org.broadinstitute.ddp.model.copy.CompositeCopyConfigurationPair;
+import org.broadinstitute.ddp.model.copy.CopyAnswerLocation;
+import org.broadinstitute.ddp.model.copy.CopyConfiguration;
+import org.broadinstitute.ddp.model.copy.CopyConfigurationPair;
+import org.broadinstitute.ddp.model.copy.CopyLocation;
+import org.broadinstitute.ddp.model.copy.CopyLocationType;
 import org.jdbi.v3.core.result.LinkedHashMapRowReducer;
 import org.jdbi.v3.core.result.RowView;
 import org.jdbi.v3.sqlobject.CreateSqlObject;
