+++ pepper-apis/src/test/java/org/broadinstitute/ddp/db/dao/CopyConfigurationDaoTest.java
@@ -19,12 +19,12 @@ import org.broadinstitute.ddp.model.activity.definition.template.Template;
 import org.broadinstitute.ddp.model.activity.types.DateFieldType;
 import org.broadinstitute.ddp.model.activity.types.DateRenderMode;
 import org.broadinstitute.ddp.model.activity.types.TextInputType;
-import org.broadinstitute.ddp.model.event.CompositeCopyConfigurationPair;
-import org.broadinstitute.ddp.model.event.CopyAnswerLocation;
-import org.broadinstitute.ddp.model.event.CopyConfiguration;
-import org.broadinstitute.ddp.model.event.CopyConfigurationPair;
-import org.broadinstitute.ddp.model.event.CopyLocation;
-import org.broadinstitute.ddp.model.event.CopyLocationType;
+import org.broadinstitute.ddp.model.copy.CompositeCopyConfigurationPair;
+import org.broadinstitute.ddp.model.copy.CopyAnswerLocation;
+import org.broadinstitute.ddp.model.copy.CopyConfiguration;
+import org.broadinstitute.ddp.model.copy.CopyConfigurationPair;
+import org.broadinstitute.ddp.model.copy.CopyLocation;
+import org.broadinstitute.ddp.model.copy.CopyLocationType;
 import org.broadinstitute.ddp.util.TestDataSetupUtil;
 import org.broadinstitute.ddp.util.TestFormActivity;
 import org.junit.BeforeClass;
